refactor: Extract schedule creation logic based on user information #35
